code/requirements.txt
opencv_python==4.5.1.48
imgaug==0.4.0
joblib==1.0.1
ConfigArgParse==1.5.2
filelock==3.0.12
tqdm==4.51.0
h5py==3.3.0
diffdist==0.1
albumentations==1.0.3
scipy==1.6.2
strconv==0.4.2
numpy==1.20.0
pandas==1.2.4
tabulate==0.8.9
fontTools==4.27.1
Pillow==8.4.0
PyYAML==6.0
scikit_learn==1.0
typer==0.4.0
